+++ .idea/.name
@@ -1,1 +1,1 @@
-menu.py
+main.py
